Use frontend-common in frontend
