+++ packages/frontend/src/components/example.component.tsx
@@ -1,6 +1,6 @@
+import { StateType } from 'frontend-common';
 import React from 'react';
 import { connect } from 'react-redux';
-import { StateType } from '../state/state.types';
 
 interface ExampleComponentProps {
     notifications: string[];
