+++ packages/frontend/src/index.tsx
@@ -1,5 +1,14 @@
+import axios from 'axios';
 import { ConnectedRouter, routerMiddleware } from 'connected-react-router';
 import 'custom-event-polyfill';
+import {
+    FrontendCommonReducer,
+    FrontendMiddleware,
+    frontendNotification,
+    listenToPlugins,
+    StateType,
+    ThunkResult,
+} from 'frontend-common';
 import { createBrowserHistory } from 'history';
 import * as log from 'loglevel';
 import React from 'react';
@@ -7,15 +16,12 @@ import 'react-app-polyfill/ie11';
 import 'react-app-polyfill/stable';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { AnyAction, applyMiddleware, compose, createStore } from 'redux';
+import { AnyAction, applyMiddleware, compose, createStore, Store } from 'redux';
 import { createLogger } from 'redux-logger';
 import thunk, { ThunkDispatch } from 'redux-thunk';
 import App from './App';
-import { configureFrontend } from './state/actions/frontend.actions';
-import FrontendMiddleware, { listenToPlugins } from './state/middleware/frontend.middleware';
-import AppReducer from './state/reducers/App.reducer';
-import { StateType } from './state/state.types';
 import './stylesheets/index.css';
+import loadMicroFrontends from './utilities/loadMicroFrontends';
 
 // Create the middleware
 const history = createBrowserHistory();
@@ -34,10 +40,28 @@ const composeEnhancers = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ||
 /* eslint-enable */
 
 // Create the store
-const store = createStore(AppReducer(history), composeEnhancers(applyMiddleware(...middleware)));
+const store = createStore(FrontendCommonReducer(history), composeEnhancers(applyMiddleware(...middleware)));
 
 // Listen to plugins
 listenToPlugins(store.dispatch);
+
+const configureFrontend = (store: Store<unknown, AnyAction>): ThunkResult<Promise<void>> => {
+    return async (dispatch) => {
+        // Request the settings file
+        await axios
+            .get('/settings.json')
+            .then((res) => {
+                const settings = res.data;
+                dispatch(frontendNotification(JSON.stringify(settings)));
+
+                // Load microfrontends
+                loadMicroFrontends.init(settings.plugins, store);
+            })
+            .catch((error) => {
+                console.log(`Frontend Error: loading settings.json: ${error.message}`);
+            });
+    };
+};
 
 // Dispatch a call to configure the frontend
 const dispatch = store.dispatch as ThunkDispatch<StateType, null, AnyAction>;
