--- packages/frontend/src/state/middleware/frontend.middleware.tsx
@@ -1,57 +1,0 @@
-import { AnyAction, Dispatch, Middleware } from 'redux';
-import { microFrontendMessageId, RegisterRouteType, RequestPluginRerenderType } from '../frontend.types';
-
-// const microFrontendMessageId = 'frontend';
-
-const broadcastToPlugins = (action: AnyAction): void => {
-    document.dispatchEvent(new CustomEvent(microFrontendMessageId, { detail: action }));
-};
-
-type microFrontendMessageType = CustomEvent<AnyAction>;
-
-export const listenToPlugins = (dispatch: Dispatch): void => {
-    document.addEventListener(microFrontendMessageId, (event) => {
-        console.log('Got: ', event);
-        const pluginMessage = event as microFrontendMessageType;
-        console.log(pluginMessage);
-
-        if (
-            pluginMessage.detail &&
-            pluginMessage.detail.type &&
-            pluginMessage.detail.type.startsWith(`${microFrontendMessageId}:api:`)
-        ) {
-            // Send message to Redux in parent app
-            switch (pluginMessage.detail.type) {
-                case RequestPluginRerenderType:
-                    // This is a message sent from parent app
-                    break;
-                case RegisterRouteType:
-                    console.log('Dispatching');
-                    dispatch(pluginMessage.detail);
-                    break;
-                default:
-                    console.warn(
-                        `Unexpected message received from plugin, not dispatched:\nevent.detail = ${JSON.stringify(
-                            pluginMessage.detail,
-                        )}`,
-                    );
-            }
-        } else {
-            console.error(
-                `Invalid message received from a plugin:\nevent.detail = ${JSON.stringify(pluginMessage.detail)}`,
-            );
-        }
-    });
-};
-
-// No need to store otherwise store => next => action
-// May need store if we need a plugin rerender
-const FrontendMiddleware: Middleware = (() => (next: Dispatch<AnyAction>) => (action: AnyAction): AnyAction => {
-    if (action.payload && action.payload.broadcast) {
-        broadcastToPlugins(action);
-    }
-
-    return next(action);
-}) as Middleware;
-
-export default FrontendMiddleware;
